requirements.txt
aiogram==2.25
aiohttp==3.8.6
aiosignal==1.4.0
async-timeout==4.0.3
attrs==25.4.0
Babel==2.9.1
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.1.4
charset-normalizer==3.4.4
frozenlist==1.8.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
propcache==0.4.1
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.22.0
